Move TranCalculator to its own header.
We need to refactor to make FieldData accessible from the simulator
for running in parallel with the TranCalculator.
