+++ CMakeLists_files.cmake
@@ -574,6 +574,7 @@ if(ENABLE_ECL_INPUT)
        opm/parser/eclipse/EclipseState/Edit/EDITNNC.hpp
        opm/parser/eclipse/EclipseState/Grid/Keywords.hpp
        opm/parser/eclipse/EclipseState/Grid/GridDims.hpp
+       opm/parser/eclipse/EclipseState/Grid/TranCalculator.hpp
        opm/parser/eclipse/EclipseState/Grid/TransMult.hpp
        opm/parser/eclipse/EclipseState/Grid/PinchMode.hpp
        opm/parser/eclipse/EclipseState/Grid/MULTREGTScanner.hpp
