+++ opm/parser/eclipse/EclipseState/Grid/FieldPropsManager.hpp
@@ -21,6 +21,7 @@
 
 #include <memory>
 #include <vector>
+#include <opm/parser/eclipse/EclipseState/Grid/TranCalculator.hpp>
 
 namespace Opm {
